Optimization
Made all models prefabs for ease of access between scenes. Did the same for Sounds. Currently in the process of removing Scenemanagers and Gamemanagers to just use DontDestroyOnLoad instead. Current problem with buttons being able to access the scenemanager script, as I can't find a way to reference dynamically. Also created CollectedItems script with an array of the food objects in scene. Makes it easier to determine, instead of "last item == null"