Optimization again
Made buttons assigner script that allows me to drag in buttons, that have an onClick listener which then reference to the Scene script to open corresponding scene. Also sorted out a lot of static variables and referencing between scripts to get everything to work. Level 1 and 2 are now completed with working food arrays and buttons etc.