Fixed: Score not saving separately for each scene
Fixed the bug where the score would save as one instead of for each scene separately. The error was that the sceneName was being declared in Start(), and due to the GameManager being a singleton that is carried through scenes, it was never updated. So even when in Level 2, the sceneName would still be declared as Level 1. To fix, just moved the code to Update().